tested files and changed fetch for a few functions

+++ .idea/database_manip/playlists.py
@@ -1,18 +1,26 @@
 from cursor import query
+import psycopg2
 
 def create_playlist(uid):
-  playlist_name = input("Enter playlist name here: ").strip()
-  while not playlist_name:
-      print("Empty playlist name")
+  try:
       playlist_name = input("Enter playlist name here: ").strip()
-  pid = query(f"INSERT INTO playlist (name) VALUES ('{playlist_name}') RETURNING pid", fetch=True)[0][0]
-  if pid is not None:
-      query(f"INSERT INTO createsp VALUES ({uid},{pid})")
-      print(f"Playlist {playlist_name} created")
-      return pid
-  else:
-      print("PID not found for playlist")
-      return None
+      if not playlist_name:
+        print("Empty playlist name")
+        return
+      pid = query(f"INSERT INTO playlist (name) VALUES ('{playlist_name}') RETURNING pid", fetch=True)[0][0]
+      if pid is not None:
+        query(f"INSERT INTO createsp VALUES ({uid},{pid})")
+        print(f"Playlist {playlist_name} created")
+        return pid
+      else:
+        print("PID not found for playlist")
+        return None
+  except  psycopg2.errors.UniqueViolation as e:
+     print("Playlists cannot have duplicate names")
+     exit
+  except Exception as e:
+     print("Database error")
+     exit
 
 def get_pid(p_name):
   p_name = p_name.strip()
@@ -23,26 +31,26 @@ def get_pid(p_name):
       return None
 
 def rename_playlist():
-  old_name = input("Enter current name of playlist here: ")
+  old_name = input("Enter current name of playlist here: ").strip()
   pid = get_pid(old_name)
-  while not pid:
+  if not pid:
        print(f"Playlist '{old_name}' does not exist")
-       old_name = input("Enter current name of playlist here: ")
-  new_name = input("Enter new name of playlist here: ")
-  while not new_name:
+       return
+  new_name = input("Enter new name of playlist here: ").strip()
+  if not new_name:
        print("Empty playlist name")
-       new_name = input("Enter current name of playlist here: ")
+       return
   query(f"UPDATE playlist SET name='{new_name}' WHERE pid={pid}")
   print(f"Playlist '{old_name}' changed to '{new_name}'")
   
        
 
 def delete_playlist():
-  name = input("Enter name of playlist to be deleted: ")
+  name = input("Enter name of playlist to be deleted: ").strip()
   pid = get_pid(name)
-  while not pid:
+  if not pid:
        print(f"Playlist '{name}' does not exist")
-       name = input("Enter name of playlist to be deleted: ")
+       return
   query(f"DELETE FROM addedto WHERE pid={pid}")
   query(f"DELETE FROM listensto WHERE pid={pid}")
   query(f"DELETE FROM playlist WHERE pid={pid}")
